apps/mp3-genre-analyzer/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
essentia-tensorflow
numpy<2
